Reverted api changes
While they worked, it was not possible to find a way to replace the offlinedbhandler with a mock while testing in time, for it to be pushed with the rest.
Therefor it has been decided to revert theses changes for now, such that everything in the running code is tested and have been proved to work.